# Resampling lookups, trailing empty coefficients, precision