--- a/Tests/test_image_resample.py
+++ b/Tests/test_image_resample.py
@@ -274,5 +274,22 @@ class CoreResamplePassesTest(PillowTestCase):
         self.assertEqual(Image.core.getcount(), count + 2)
 
 
+class CoreResampleCoefficientsTest(PillowTestCase):
+    def test_reduce(self):
+        test_color = 254
+        # print ''
+
+        for size in range(400000, 400010, 2):
+            # print '\r', size,
+            i = Image.new('L', (size, 1), 0)
+            draw = ImageDraw.Draw(i)
+            draw.rectangle((0, 0, i.size[0] // 2 - 1, 0), test_color)
+
+            px = i.resize((5, i.size[1]), Image.BICUBIC).load()
+            if px[2, 0] != test_color // 2:
+                self.assertEqual(test_color // 2, px[2, 0])
+                # print '\r>', size, test_color // 2, px[2, 0]
+
+
 if __name__ == '__main__':
     unittest.main()