@homm 2016-07-03
Failed tests are not related to resizing at all. There are two bugs: in PyAccess module on CPython and in PyAccess tests. I've fixed tests (PR impending), looking at the module.