@homm 2016-07-03
Now green. Fixes in https://github.com/python-pillow/Pillow/pull/2009