@wiredfool 2016-08-25
Merged in #2091